backend/requirements.txt
django==5.2.1
djangorestframework==3.16.0
python-slugify==8.0.4
django-cors-headers==4.7.0
django-filter